New Version of louvain working :) deleted old
