--- src/es/upc/fib/prop/shared13/CommunityAlgorithm.java
@@ -1,14 +1,0 @@
-package es.upc.fib.prop.shared13;
-
-import es.upc.fib.prop.shared13.louvain.LouvainGraph;
-
-/**
- * Created by ondrej on 15.4.15.
- */
-public abstract class CommunityAlgorithm {
-
-
-
-
-
-}
