--- src/es/upc/fib/prop/shared13/CommunitySet.java
@@ -1,8 +1,0 @@
-package es.upc.fib.prop.shared13;
-
-/**
- * Created by miquel on 28/04/15.
- */
-public class CommunitySet
-{
-}
