--- src/es/upc/fib/prop/shared13/louvain/Partition.java
@@ -1,11 +1,0 @@
-package es.upc.fib.prop.shared13.louvain;
-
-import java.util.LinkedHashMap;
-
-/**
- * Created by miquel on 25/04/15.
- */
-public class Partition<Node,Integer> extends LinkedHashMap<Node,Integer>
-{
-    //You need to map each node to a number which indicates its community
-}
